+++ financeiro/src/br/com/grupoferraz/financeiro/dao/RateioDAO.java
@@ -156,7 +156,7 @@ public class RateioDAO {
 				grupo.setNome(rs.getString("nome"));
 				grupo.setAtividade(rs.getString("atividade"));
 				grupo.setCrcontabil(rs.getString("crcontabil"));
-				grupo.setPeso(rs.getFloat("peso"));
+				grupo.setPeso(rs.getBigDecimal("peso"));
 				grupo.setGrupocentroresultados_codigo(rs.getInt("grupocentroresultados_codigo"));
 			}
 			return grupo;
